(MAINT) Default value for reporting
We'll want to set the ENV value to the final lumogon.com domain but this
change means that running report will report and scan will scan, and
that you can still override the default value with a flag on the CLI or
now also by passing an environment variable.

A few things to note:

I'm not using Viper here because it's currently incorrectly configured.
I also had a fight wit cobra/pflags where default values for flags actually
exist in commands they are not set against, so several things that
looked nicer don't actually work.