(MAINT) Remove viper configuration options
These don't currently work as intended due to us not using the
viper.Get* methods in the rest of the application. We're also using the
wrong call for the environment variable support BindEnv should be
SetEnvPrefix based on the comment.

Having full local and environment variable config sounds like a good
thing for the future, but given we don't actually have it now removing
this code feels the most sensible options.