(TC-131) add CopyFromContainer to dockeradapter
Adds a CopyFrom interface to the dockeradapter package, consists of a
single CopyFromContainer function which wraps the similarly named
Docker API call.

It adds a new flag, followSymlink, which if set will return a Reader for
the linked file.

As an FYI, the file returned from the Docker API Reader is a tar file.