(TC-131) add trim test and fix log messages
Adds a test to ensure we can extract the OS ID when its surrounded by
single or double quotes and also leading or trailing spaces.

Also tweaks the log messages to print strings pulled from os-release as
double-quoted strings safely escaped with Go syntax. This is so that it
will be easier to pick up escape characters in future.