(MAINT) use WaitTimeout contexts to timeout stuck harvesters
Stips out all report generation from the collector to keep it simpler
(pushed to the storage package).

Uses context.WithTimeout to allow cancellation of the collector and
attached harvesters which may block on waiting for results from the
harvesters (collector) or the rpcreceiver (attached harvester).

If these are cancelled any results already returned will be sent to the
storage backend for the creation and storage (bad name? either send to
the DB or stdout).

**NOTE** Any harvesters which have failed to return a result to the
rpcreceiver currently do not result in that harvester container being
removed.