(MAINT) Fix local container detection on linux
Prior to this commit the detection of the Lumogon scheduler container id
wasn't working for Docker Engines running on a Linux host.

This would result in the Lumogon scheduler container showing up in the
generated report.

The client was expecting the `/proc/self/cgroup` cpu entry to exist on
its own line but on some linux hosts you can have multiple cpu resources
on a single line for example:
```
4:cpu,cpuacct:/docker/c7acb1eb27bb163d07c2ce0a3e7deb36a698824472f21074662e605bf57c7521

```
This commit relaxes the regex used to extract the containerID,
adds/updates some tests and adds an extra logging statement which is
printed when the GetLocalContainerID returns an error