(MAINT) Tighten FilterDockerStream error handling
Adds tests for cases where reading from the docker stream buffer returns
an error and fixes the FilterDockerStream in cases where attempting to
read a payload after the header read has thrown an error was panicing.

Adds some more information to the readFramePayload and readFrameHeader
function comments to describe how errors are handled.