(FEAT) Add container filesystem size capability
This is currently only close to working (or so I think). It adds a new
"filesystem" capability, uses `ContainerList` with a container `id`
filter to retrieve information about our specific container, and then
extracts the `SizeRW` and `SizeRootFs` data.

The only downside is that these two values are always 0 so far.