(TC-157) Output message when scanning each container
Based on some of the user feedback, the fact we don't output anything
when running a scan can cause some users to think something is amiss.
This commit adds a single message to the output for each container
scanned when running scan or report. A few considerations:

* The output goes to STDERR, with the intention that this is "diagnostic
  information" rather than output which would be usefully parsed. This
  maintains the ability to treat the actual output as JSON with JQ or
  other tooling.
* The log prefix has been changed to be Sentence case, to match the
  other tags. ie. it is not [Lumogon] rather than [lumogon].
* The logging interface was a touch confusing, in that logging.Stdout
  actually went to stderr. I've made this more abstract, ie. with Info
  and Debug.

I've purposefully not tried to delve into changing the entire logging
implementation at this point.