(MAINT) cache test nginx image
This commit caches the test nginx image to disk and loads it locally
rather than having to pull the image from within each DIND container.

This speeds up the execution as you're only pulling in from the network
once.

It also sidesteps the fact that Docker 1.5.0 doesn't work with the Hub
so would fail to pull the image.