(DI-489) (TC-155) Identify API version when creating client
This commit updates the dockeradapters New function to detect the API
version of the Docker server to which its connecting before creating the
client that Lumogon will use for harvesting.

It first attempts to connect without specifying an API version where it
determines the Server API version and its ID. A new client connection is
then established using the version returned from the server with that
version also stored in the concreteDockerClient type.

Note that this still only supports socket connections, but extending to
TCP connections should now be relatively straight forward.

The MinSupportedAPIVersion is set to `1.21` which corresponds to Docker
Engine `1.9.x`.

Support for earlier servers require being able to identify the scheduler
container (with API < 1.21), and an alternative mechanism to detect
target container OS (with API < 1.20)