(MAINT) Improve opensuse rpm package support
Prior to this commit the rpm capability was reliant on the presence of
the `rpm` binary to list installed packages. This is not present on
opensuse images, however the `rpmquery` command is present.

This commit adds a fallback to `rpmquery` when enumerating insalled
packages.

*note* it also strips the `-(none)` returned from some packages which
have the arch defined as none rather than an explicit `noarch`.